joh/administrative koala (#181514)
* :lipstick: work with Markdown/EditResponse sooner

* move session object and friends into separete service
